# apps/backend/internal/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	CookieAuthScopes cookieAuthContextKey = "cookieAuth.Scopes"
)

// AuthRequest defines model for AuthRequest.
type AuthRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// AuthResponse defines model for AuthResponse.
type AuthResponse struct {
	User User `json:"user"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Message Human-readable error message
	Message string `json:"message"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// Scene defines model for Scene.
type Scene struct {
	CreatedAt time.Time `json:"createdAt"`

	// Data Scene content as stored by the frontend
	Data SceneData `json:"data"`

	// Id Unique identifier (UUID)
	Id string `json:"id"`

	// Name Human-readable scene name
	Name string `json:"name"`

	// ShareToken Opaque share token. Non-null means the scene is publicly accessible via GET /api/v1/share/{token}. Null means sharing is disabled.
	ShareToken *string   `json:"share_token,omitempty"`
	UpdatedAt  time.Time `json:"updatedAt"`
}

// SceneData Scene content as stored by the frontend
type SceneData struct {
	// Groups Item groups
	Groups []SceneGroup `json:"groups"`

	// Items Scene items (furniture, panels, etc.)
	Items []SceneItem `json:"items"`

	// Room Room dimensions in mm. If absent, defaults are used (4000x4000x3000).
	Room *struct {
		Depth  int `json:"depth"`
		Height int `json:"height"`
		Width  int `json:"width"`
	} `json:"room,omitempty"`

	// Version Snapshot format version
	Version int `json:"version"`
}

// SceneGroup Opaque scene group — shape defined by frontend SceneGroup type
type SceneGroup map[string]interface{}

// SceneInput defines model for SceneInput.
type SceneInput struct {
	// Data Scene content as stored by the frontend
	Data SceneData `json:"data"`

	// Name Human-readable scene name
	Name string `json:"name"`
}

// SceneItem Opaque scene item — shape defined by frontend SceneItem type
type SceneItem map[string]interface{}

// SceneRenameInput defines model for SceneRenameInput.
type SceneRenameInput struct {
	// Name New scene name
	Name string `json:"name"`
}

// SceneSummary defines model for SceneSummary.
type SceneSummary struct {
	CreatedAt time.Time `json:"createdAt"`

	// Id Unique identifier (UUID)
	Id string `json:"id"`

	// Name Human-readable scene name
	Name string `json:"name"`

	// ShareToken Opaque share token. Non-null means the scene is publicly accessible via GET /api/v1/share/{token}. Null means sharing is disabled.
	ShareToken *string   `json:"share_token,omitempty"`
	UpdatedAt  time.Time `json:"updatedAt"`
}

// User defines model for User.
type User struct {
	Email openapi_types.Email `json:"email"`

	// Id User ID (UUID)
	Id string `json:"id"`
}

// SceneId defines model for SceneId.
type SceneId = string

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// cookieAuthContextKey is the context key for cookieAuth security scheme
type cookieAuthContextKey string

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = AuthRequest

// RegisterJSONRequestBody defines body for Register for application/json ContentType.
type RegisterJSONRequestBody = AuthRequest

// CreateSceneJSONRequestBody defines body for CreateScene for application/json ContentType.
type CreateSceneJSONRequestBody = SceneInput

// RenameSceneJSONRequestBody defines body for RenameScene for application/json ContentType.
type RenameSceneJSONRequestBody = SceneRenameInput

// UpdateSceneJSONRequestBody defines body for UpdateScene for application/json ContentType.
type UpdateSceneJSONRequestBody = SceneInput