# apps/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RoomTool API
  version: "1.0.0"
  description: REST API for RoomTool wardrobe designer
paths:
...
  /api/v1/scenes/{id}/share:
    parameters:
      - $ref: "#/components/parameters/SceneId"
    post:
      operationId: EnableShare
      summary: Enable sharing (idempotent)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Share token (existing or newly created)
          content:
            application/json:
              schema:
                type: object
                required: [share_token]
                properties:
                  share_token:
                    type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: DisableShare
      summary: Revoke sharing
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Sharing revoked
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    SceneId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
  schemas:
...
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: Human-readable error message